+++ gnucash_importer/read_entry.py
@@ -1,25 +1,28 @@
 # -*- coding: utf-8 -*-
-
-"""
-Module to read all file formats that will be supported
-"""
+"""Module to read all file formats that will be supported."""
+from abc import abstractmethod
 import logging
-from ofxparse import OfxParser  # ofxparse 0.14+
+from ofxparse import OfxParser  # ofxparse +0.14
 
 class EntryReader(object):
+    """Aggreates all functions and act as interface."""
     transactions = None
 
     def __init__(self):
         pass
 
+    @abstractmethod
     def get_transaction(self, report_file):
+        """Given a file, get all transactions there."""
         pass
 
     def print_transactions(self):
+        """Print to standard output all transactions stored in this class. Require get_transaction before use this method."""
         for transaction in self.transactions:
             self.print_transaction(transaction)
     
     def print_transaction(self, transaction):
+        """Print to standard output one transaction at time."""
         print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
         print("amount....: %s" % (transaction.amount))
         print("checknum..: %s" % (transaction.checknum))
@@ -32,7 +35,9 @@ class EntryReader(object):
         print("type......: %s" % (transaction.type))
 
 class OfxReader(EntryReader):
+    """Implement EntryReader class specialized in OFX data file."""
     def get_transactions(self, report_file):
+        """Implement super behavior."""
         report = open(report_file)
         ofx = OfxParser.parse(report)
         self.transactions = ofx.account.statement.transactions
@@ -41,9 +46,13 @@ class OfxReader(EntryReader):
         return self.transactions
 
 class QifReader(EntryReader):
+    """Implement EntryReader class specialized in QIF data file."""
     def get_transactions(self, report_file):
-        print("TODO stub method")
+        """TODO STUB METHOD. Implement super behavior."""
+        print("TODO STUB METHOD.")
 
 class CsvReader(EntryReader):
+    """Implement EntryReader class specialized in CSV data file."""
     def get_transactions(self, report_file):
-        print("TODO stub method")
+        """TODO STUB METHOD. Implement super behavior."""
+        print("TODO STUB METHOD.")