+++ gnucash_importer/util.py
@@ -1,10 +1,16 @@
 # -*- coding: utf-8 -*-
+"""
+Util module to concetrate all reusable code.
+
+This modules support all other classes/modules with usefull source code that doesn't belong to any other entity.
+"""
 
 import configparser
 import logging
 from termcolor import colored, cprint
 
 class Util:
+    """Helper class used to provide configuration, defaults and so on."""
     config = configparser.ConfigParser()
     LOG_FORMAT_FULL = colored('[%(asctime)s][%(process)d:%(processName)s]', 'green', attrs=['bold', 'dark']) + colored('[%(filename)s#%(funcName)s:%(lineno)d]', 'white', attrs=['bold', 'dark']) + colored('[%(levelname)s]', 'magenta', attrs=['bold', 'dark']) + ' %(message)s'
     LOG_FORMAT_DEBUG = colored('[%(filename)s#%(funcName)s:%(lineno)d]', 'white', attrs=['bold', 'dark']) + colored('[%(levelname)s]', 'magenta', attrs=['bold', 'dark']) + ' %(message)s'
@@ -26,6 +32,7 @@ class Util:
     DEFAULT_BRADESCO_SAVINGS_FROM = None
 
     def __init__(self):
+        """Grab all configuration in setup.cfg"""
         self.config.read('setup.cfg')
         self.DEFAULT_CURRENCY = self.config['app_init']['default_currency']
         self.DEFAULT_GNUCASH_FILE = self.config['app_init']['default_gnucash_file']
@@ -43,10 +50,8 @@ class Util:
         self.DEFAULT_BRADESCO_SAVINGS_TO = self.config['app_init']['default_bradesco_savings_to']
         self.DEFAULT_BRADESCO_SAVINGS_FROM = self.config['app_init']['default_bradesco_savings_from']
 
-    """
-    Helper to discovery API of an object
-    """
     def show_methods(obj):
+        """Helper to discovery API of an object."""
         logging.basicConfig(level = logging.DEBUG, format = Util.LOG_FORMAT_DEBUG)
         logging.debug(colored("_______________________________________________________", 'cyan'))
         logging.debug(colored("type(obj)..: {t}".format(t = type(obj)), 'cyan'))
@@ -58,13 +63,17 @@ class Util:
         logging.debug(colored("_______________________________________________________", 'cyan'))
 
     def info(msg):
+        """This function standardize the message and simplified the use to standard output."""
         return colored(msg, 'blue')
 
     def warning(msg):
+        """This function standardize the message and simplified the use to standard output."""
         return colored(msg, 'yellow', attrs=['bold'])
 
     def error(msg):
+        """This function standardize the message and simplified the use to standard output."""
         return colored(msg, 'red', attrs=['bold', 'underline'])
 
     def debug(msg):
+        """This function standardize the message and simplified the use to standard output."""
         return colored(msg, 'grey', attrs=['reverse', 'bold', 'underline'])